backend/requirements.txt
django
coverage
pytest
pytest-django 
pytest-cov
pytest-pythonpath
codecov
djangorestframework
markdown
django-filter
pillow
